:art: Refactor project structure for clarity and best practices
- Move business and API step definitions to clear domain packages
- Organize feature files into api/ and business/ folders
- Move business service classes and tests to business/service
- Remove unused files, empty folders, and example Main.java
- Update runners, imports, and README with new structure
- Add Mermaid diagram for visual architecture
