+++ src/main/java/com/example/business/service/SimpleEmailValidationStrategy.java
@@ -1,4 +1,4 @@
-package com.example.service;
+package com.example.business.service;
 
 import com.example.model.User;
