+++ src/test/java/com/example/business/runners/RegisterUserTestRunner.java
@@ -1,4 +1,4 @@
-package runners;
+package com.example.business.runners;
 
 import io.cucumber.junit.CucumberOptions;
 import net.serenitybdd.cucumber.CucumberWithSerenity;
@@ -6,8 +6,8 @@ import org.junit.runner.RunWith;
 
 @RunWith(CucumberWithSerenity.class)
 @CucumberOptions(
-        features = "src/test/resources/features/register_users.feature",
-        glue = "com.example.stepdefinitions",
+        features = "src/test/resources/features/business/register_users.feature",
+        glue = "com.example.business.steps",
         plugin = {"pretty"},
         snippets = CucumberOptions.SnippetType.CAMELCASE
 )
